Analytics cutover
