+++ analytics/src/analytics/logs/ecs_background_task.py
@@ -68,7 +68,7 @@ def ecs_background_task(task_name: str) -> Callable[[Callable[P, T]], Callable[P
 
 
 @contextlib.contextmanager
-def _ecs_background_task_impl(task_name: str) -> Generator[None, None, None]:
+def _ecs_background_task_impl(task_name: str) -> Generator[None]:
     # The actual implementation, see the docs on the
     # decorator method above for details on usage
