+++ analytics/src/analytics/logs/formatters.py
@@ -11,7 +11,7 @@ See https://docs.python.org/3/library/logging.html#formatter-objects
 import json
 import logging
 from collections.abc import Callable
-from datetime import date, datetime, timezone
+from datetime import UTC, date, datetime
 from decimal import Decimal
 from enum import Enum
 from typing import Any, TypeVar
@@ -106,7 +106,7 @@ class HumanReadableFormatter(logging.Formatter):
         """Format the log record to be human-readable with color-coding."""
         message = super().format(record)
         return decodelog.format_line(
-            datetime.fromtimestamp(record.created, tz=timezone.utc),
+            datetime.fromtimestamp(record.created, tz=UTC),
             record.name,
             record.funcName,
             record.levelname,
